tablas de historial

+++ src/main/java/com/stoqing/reservas/entities/dto/Audit.java
@@ -23,7 +23,7 @@ public class Audit{
     private LocalDateTime deletedAt;
 
     @Column(name="modified_by")
-    private Long modifiedBy;
+    private Integer modifiedBy;
 
     @PrePersist
     public void prePersist() {
